+++ vm_windows.tf
@@ -1,12 +1,13 @@
 resource "azurerm_windows_virtual_machine" "machine" {
-  count                 = local.windows_vm
+  for_each              = var.source_image_reference.offer == "WindowsServer" ? { Windows = "true" } : {}
   name                  = var.name
-  resource_group_name   = data.azurerm_resource_group.rg.name
-  location              = var.location != null ? var.location : data.azurerm_resource_group.rg.location
+  resource_group_name   = var.resource_group
+  location              = var.location
   size                  = var.vm_size
   admin_username        = var.admin_user.username
   admin_password        = var.admin_user.password
-  network_interface_ids = var.network_interface_ids
+  network_interface_ids = local.network_interface_ids
+
   os_disk {
     caching              = var.os_disk.caching
     storage_account_type = var.os_disk.storage_account_type