+++ CHANGELOG.md
@@ -1,9 +1,16 @@
 # Changelog
+
+## v1.2.0
+- #4 Add the possibility to create a network interface through the module
+- #10 Fix how the module decides what resource to use
+- #11 Remove data.azurerm_resource_group, due to unintentional side-effects
+
 ## v1.1.0
 - Added Active Directory domain join
 
 ## v1.0.1
 Bug fix, logic related to location was looking for the wrong data type.
+
 ## v1.0.0
 Initial version of the module, ready for testing in production.